requirements.txt
tensorflow-cpu
flask
numpy
opencv-python-headless
pillow
matplotlib
scikit-learn

